rename db.methods to controller
